+++ tests/RuleKit.Tests/Models/DateStringModel.cs
@@ -1,0 +1,11 @@
+﻿namespace RuleKit.Tests.Models
+{
+    internal sealed class DateStringModel
+    {
+        /// <summary>
+        /// Gets or sets the value to validate.
+        /// </summary>
+        [DateString("yyyyMMdd")]
+        public string? Value { get; set; }
+    }
+}